\name{DeconUtils-package}
\alias{DeconUtils-package}
\alias{DeconUtils}
\docType{package}
\title{
\packageTitle{DeconUtils}
}
\description{
\packageDescription{DeconUtils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DeconUtils}
\packageIndices{DeconUtils}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{DeconUtils}

Maintainer: \packageMaintainer{DeconUtils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
